Fix volatile reprocessing
