+++ setup.py
@@ -6,7 +6,7 @@ setup(name='data_processing',
       author='Finnish Meteorological Institute',
       license='MIT License',
       install_requires=['cloudnetpy>=1.18.0',
-                        'cloudnetme',
+                        'cloudnetme>=0.1.1',
                         'pytest',
                         'pylint',
                         'requests',
